requirements.txt
phidata
google-generativeai
fastapi
uvicorn
chdb>=2.0
